Update GUI to use PyQt5 and Python 3
May also work with PyQt4 and Python 2, but further testing needs to be
conducted. Also added and tested additional functionality and changed
some instance variable names. Lastly, made the program prioritize XML
files when autoloading. The default save still needs to be changed to
'txt', with an option to 'export to txt'.